+++ src/lib/i18n/dictionaries/dictionary.json
@@ -3068,8 +3068,6 @@
   "magicMushroom.description": "Used to cook advanced recipes",
   "wildMushroom.description": "Used to cook basic recipes",
   "honey.description": "Used to sweeten your cooking",
-  "restock.one.buck": "You are going to use 20 Gems to restock all shop items in the game.",
-  "restock.sure": "Are you sure you want to Restock?",
   "restock.tooManySeeds": "You have too many seeds in your basket!",
   "seeds.reachingInventoryLimit": "You are reaching your seed basket limit!",
   "crops.noCropsToSell": "You have no {{cropName}} to sell!",
@@ -3740,8 +3738,13 @@
   "crafting.readySoon": "{{name}} will be ready soon...",
   "crafting.expansionSoon": "Your land will be ready soon...",
   "crafting.viewRecipes": "View recipes",
-  "gems.buyReplenish": "Would you like to replenish the shops now for 20 Gems?",
-  "gems.buyReplenish.enhanced": "Would you like to replenish the {{shopName}} now for {{gemPrice}} Gems?",
+  "gems.buyReplenish": "Would you like to restock the shops for 20 Gems?",
+  "gems.buyReplenish.enhanced": "Would you like to restock the {{shopName}} for {{gemPrice}} Gems?",
+  "restock.category": "Restock {{category}}",
+  "restock.all": "Restock all",
+  "market": "market",
+  "workbench": "workbench",
+  "treasure.shop": "treasure shop",
   "gems.restockToMaxStock": "Restocks up to maximum stock",
   "gems.noShipment": "No Shipment",
   "gems.buyStock": "Please buy your stock to get your shipment.",
